v0.14  commit

+++ python/paddle/fluid/tests/unittests/test_mine_hard_examples_op.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import print_function
-
 import unittest
 import numpy as np
 import sys
@@ -59,13 +57,13 @@ class TestMineHardExamplesOp(OpTest):
         self.sample_size = 0
         self.mining_type = "max_negative"
         self.cls_loss = np.array([[0.1, 0.1, 0.3],
-                                  [0.3, 0.1, 0.1]]).astype('float64')
+                                  [0.3, 0.1, 0.1]]).astype('float32')
 
         self.loc_loss = np.array([[0.1, 0.2, 0.3],
-                                  [0.3, 0.4, 0.1]]).astype('float64')
+                                  [0.3, 0.4, 0.1]]).astype('float32')
 
         self.match_dis = np.array([[0.2, 0.4, 0.8],
-                                   [0.1, 0.9, 0.3]]).astype('float64')
+                                   [0.1, 0.9, 0.3]]).astype('float32')
 
         self.match_indices = np.array([[0, -1, -1],
                                        [-1, 0, -1]]).astype('int32')
@@ -83,10 +81,10 @@ class TestMineHardExamplesOpHardExample(TestMineHardExamplesOp):
         self.sample_size = 2
 
         self.cls_loss = np.array([[0.5, 0.1, 0.3],
-                                  [0.3, 0.1, 0.1]]).astype('float64')
+                                  [0.3, 0.1, 0.1]]).astype('float32')
 
         self.loc_loss = np.array([[0.2, 0.2, 0.3],
-                                  [0.3, 0.1, 0.2]]).astype('float64')
+                                  [0.3, 0.1, 0.2]]).astype('float32')
 
         self.match_indices = np.array([[0, -1, -1],
                                        [-1, 0, -1]]).astype('int32')
